Implement Sampling OTEP
