+++ api/lib/opentelemetry/trace.rb
@@ -42,7 +42,6 @@ module OpenTelemetry
 end
 
 require 'opentelemetry/trace/trace_flags'
-require 'opentelemetry/trace/samplers'
 require 'opentelemetry/trace/span_context'
 require 'opentelemetry/trace/span_kind'
 require 'opentelemetry/trace/span'
