+++ sdk/lib/opentelemetry/sdk/internal.rb
@@ -6,8 +6,15 @@
 
 module OpenTelemetry
   module SDK
+    # @api private
+    #
     # Internal contains helpers used by the reference implementation.
     module Internal
+      extend self
+
+      def boolean?(value)
+        value.is_a?(TrueClass) || value.is_a?(FalseClass)
+      end
     end
   end
 end
